Refactor according to conventions
