+++ site/src/components/AdminDropdown/AdminDropdown.tsx
@@ -3,12 +3,12 @@ import ListItemText from "@material-ui/core/ListItemText"
 import { fade, makeStyles, Theme } from "@material-ui/core/styles"
 import AdminIcon from "@material-ui/icons/SettingsOutlined"
 import React, { useState } from "react"
-import { navHeight } from "../../../theme/constants"
-import { BorderedMenu } from "../../BorderedMenu"
-import { BorderedMenuRow } from "../../BorderedMenu/BorderedMenuRow"
-import { BuildingIcon } from "../../Icons/BuildingIcon"
-import { UsersOutlinedIcon } from "../../Icons/UsersOutlinedIcon"
-import { CloseDropdown, OpenDropdown } from "../Arrows"
+import { navHeight } from "../../theme/constants"
+import { BorderedMenu } from "../BorderedMenu/BorderedMenu"
+import { BorderedMenuRow } from "../BorderedMenuRow/BorderedMenuRow"
+import { CloseDropdown, OpenDropdown } from "../DropdownArrows/DropdownArrows"
+import { BuildingIcon } from "../Icons/BuildingIcon"
+import { UsersOutlinedIcon } from "../Icons/UsersOutlinedIcon"
 
 export const Language = {
   menuTitle: "Admin",
